refactor(webview-ext): introduce shared entry (wire types, TypedEventSink, BaseRouterContext)
