+++ packages/vscode-ext-webview/src/shared/TypedEventSink.ts
